requirements.txt
asgiref==3.11.0
crispy-bootstrap4==2025.6
Django==5.2.8
django-crispy-forms==2.5
django-environ==0.12.0
gunicorn==23.0.0
packaging==25.0
pillow==12.0.0
sqlparse==0.5.3
